+++ src/clients/browser/spells.js
@@ -1,0 +1,3 @@
+import { spellWatcher } from "~/spells";
+
+spellWatcher();